Move golang dir
