+++ golang/addons/addons_test.go
