+++ golang/addons/template/grafana-service.yaml
