+++ golang/addons/template/heapster-controller.template
