+++ golang/addons/template/heapster-service.yaml
