+++ golang/addons/template/influxdb-grafana-controller.template
