+++ golang/addons/template/influxdb-service.yaml
