+++ golang/addons/template/skydns-service.template
