+++ golang/addons/template/skydns.template
