+++ golang/cloud-config-server/build.sh
