+++ golang/cloud-config-server/server_test.go
