+++ golang/clusterdesc/config_test.go
