+++ golang/clusterdesc/master_test.go
