+++ golang/template/cluster-desc.sample.yaml
