+++ golang/template/template.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 	"text/template"
 
-	"github.com/k8sp/sextant/cloud-config-server/certgen"
-	"github.com/k8sp/sextant/cloud-config-server/clusterdesc"
+	"github.com/k8sp/sextant/golang/certgen"
+	"github.com/k8sp/sextant/golang/clusterdesc"
 	"github.com/topicai/candy"
 	"gopkg.in/yaml.v2"
 )
