+++ golang/template/templatefiles/cc-centos.template
