+++ golang/template/templatefiles/cc-common.template
