+++ golang/template/templatefiles/cc-coreos.template
