+++ golang/template/templatefiles/cloud-config.template
