+++ golang/validate-yaml/validate-yaml.go
@@ -13,9 +13,9 @@ import (
 	"errors"
 	"flag"
 	"github.com/golang/glog"
-	"github.com/k8sp/sextant/cloud-config-server/certgen"
-	"github.com/k8sp/sextant/cloud-config-server/clusterdesc"
-	cctemplate "github.com/k8sp/sextant/cloud-config-server/template"
+	"github.com/k8sp/sextant/golang/certgen"
+	"github.com/k8sp/sextant/golang/clusterdesc"
+	cctemplate "github.com/k8sp/sextant/golang/template"
 	"github.com/topicai/candy"
 	"gopkg.in/yaml.v2"
 	"io/ioutil"
